# Change linked-list of timers into min binary heap.