Based on the [A* min heap code](#1236). This will allow us some
extra room working with heavy use of timers, and also provide
an easier way to change timer values if we need it (compared to
delete/create new).

#### Commit 1
Timers are currently kept in a linked list, unfortunately the timers
generally have a cluster at the beginning of the list (mob movement)
so every insert/delete has to traverse minimum 30-100 elements before
potentially finding the correct place to modify.

Changing to min-heap means that worst-case scenario, 10 timers have
to be looked at (generally it is much better than this).

On my system, CreateTimer drops from 1us to 0.350us, DeleteTimer
from 1.1us to 0.6us and GetTimeRemaining from 8.6us to 1.5us.

#### Commit 2
Use a unnamed struct/anonymous union in timer_node to allow moving
data faster in the TimerSwapIndex function. I'm not 100% sure why
this is faster, can only think that the compiler can't optimise the 12 swaps
necessary to do it the other way. Moving the whole struct followed by
reassigning heap_index is comparable in speed to the 12 swaps.

I don't know if this is "good C/C++" or if it will even compile on other
compilers, so I put this part in a separate commit if I need to remove it.

#### Commit 3
Add comments to function definitions and put heap-related code
into a region.

#### Commit 4
Timer heap memory will now be expanded if it hits the maximum allocated
timer nodes (default 2000).

Added a new menu option to the admin menu, Check, which currently has
one option, timerheap. "Check timerheap" will traverse the timer heap
and check it for validity (children of a node larger than parent). The
result and time to run is displayed in admin window/server interface.